Make the macOS bundle compatible osx 10.12
On macOS 10.12.6 (as far as I can test), running `make app` creates a
bundle that refuses to start (it complain application is only compatible
with more recent versions of macOS. The binary works great when running
it directly.

As far as I can tell this fixes the issue and the bundle works great
(despite the other issues reported for login shell, but I guess that's
another story).

I'm going to _assume_ that targeting 10.11 or lower would also work.
I don't know what effect this have on the final binary and whether this
removed  some optimisation. Maybe that could be conditionally setup
depending on the version of macOS this is built on?