Set macOS target in Makefile.
This attempt to extract the current macos target in the makefile, and
should allow the bundle to work on the machine this was build on.

One of the advantage is that it will work on old os, by still allowing
further optimisation on recent ones.

This has one big drawback vs build.rs. Make will _not_ detect that
cargo-build need to be re-ran if it has been ran before.
That is to say the two command in sequence will lead to a broken bundle.
 - cargo build --release
 - make app