Change default URL launchers
The default configuration has been altered to use `xdg-open` to launch
URLs on linux and `open` to do so on macOS. This should allow clicking
on links by default on most systems.

Some changes which were made to the reverse grid iterator have also been
reverted.

Alacritty also doesn't crash anymore when the program specified as
`url_launcher` isn't installed on the system. Instead it will print a
warning to the log.