Fix URL launcher with active selection
With an active selection, the URL launcher would still open the URL when
clicking on it, instead of just clearing the selection. Using a
double-click the URL has also been launched twice.

Both these issues have been resolved by changing the `last_press_pos`
into a URL-specific `block_url_launcher` variable which only allows the
URL launcher to start when all required parameters are met.

The user can now also move his mouse during selection, but only until
the motion is big enough to start a selection. So either when the cell
is changed or when the mouse moves in between the left and right cell
side.

This change also restructures the code to be a bit more readable and
less over the place by moving the URL selection code in the `input.rs`
file into a separate method.