Build binary inside 18.04 docker image
Instead of using the 14.04 image which travis provides by default, an
18.04 docker image is now used to build the output binaries for Linux.
This affects both the .deb and a new .tar.gz binary.

The advantage of this is that while binaries compiled on 14.04, do not
work on 18.04, it does work the other way around. The generated .tar.gz
binary has been tested on 18.04, Debian, Fedora and Archlinux and all
systems were able to run it without any warnings or errors.