Replace sysout/err prints with logging
To make sure that all error and information reporting to the user is
unified, all instances of `print!`, `eprint!`, `println!` and
`eprintln!` have been removed and replaced by logging.

When `RUST_LOG` is not specified, the default Alacritty logger now also
prints to both the stdout and a log file. The log file is only created
when a message is written to it and its name is printed to stdout the
first time it is used.