Improve URL detection
Since the initial version of URL detection had some issues with being a
bit over-eager to mark things as URLs, a different approach has been
taken instead.

To make parsing easier, instead of trying to detect the URL by using a
streaming parser to the left and right of the click position, the
complete buffer is now converted to a string and then searched for
links.

With this change, it was made possible to use the `linkify` library
which handles locating the URLs in the terminal buffer text. Since it
now automatically finds all present URLs, this also brings us a step
closer to #1618.

This fixes #1753 and fixes #1727. Even though the protocols have not
been limited, only protocols using the `scheme://` format are accepted
now, which removes the root cause of #1727.