feature: report Windows panics through MessageBox.
This commit introduces a new module called `platform`, which provides
operating system and platform-level integration capability.  For now, it
targets Windows only.  It provides a panic handler that exposes such
events through the classical Win32 MessageBox facility, whereupon a user
can easily copy-and-paste the error for later forensic analysis and bug
reporting..

Why is this proposed?  In Issue #1818, I remarked that Alacritty in
Windows crashed unexpectedly since winpty was unable to find
`winpty-agent.exe`.  This error message was opaque and invisible, for I
had not started Alacritty itself from another terminal.  I reason that
few users of Alacritty on Windows would ever start a terminal from
another terminal, so it would be best to take a platform-idiomatic
approach of exposing the dialog box.