Normalize log message strings.
The general style for errors, warnings and info messages is to start
with a capitalized letter and end without a period. The main exception
is when dealing with nouns that are clearer with special case handling,
e.g. "macOS failed to work" or "ioctl is borked".

Trace and debug messages have been normalized to lowercase starts, but I
suspect we don't care to keep these as consistent.