Extract max zerowidth characters into variable
The maximum amount of zerowidth characters per cell have been extracted
into the `MAX_ZEROWIDTH_CHARS` constant in the `src/term/cell.rs` file.

This should not only make it easier to change the maximum amount of
zerowidth characters in the future, but also provide an easy way to test
Alacritty's behavior with different amounts of zerowidth characters.