Remove ubuntu perf install script
Since `perf` is a standard tool which is available in the official
repositories for most distributions, it doesn't make much sense to
provide an installation script specifically for Ubuntu. As a result of
this, the script has been removed.

This also fixes a minor typo in the `README.md` file.