Disable live-config-reload for opacity
It is now not possible anymore to update the background opacity of
alacritty while it is running.

This is not just the removal of a feature though, there have been a lot
of issues with the `with_transparency(true)` call in the `window.rs`
module. Some of these issues are flickering on resize and alacritty
crashing on startup.

To fix these issues the transparency of alacritty is now set at startup,
if alacritty is started with opacity `1.0` the window is crated
with `.with_transparency(false)`. Otherwise transparency is disabled.

Even though this definitely is not a great solution, this should be
beneficial for most users and solve a lot of annoyances that have
been mentioned in the past.

This fixes #1268.