Extract line calculations in separate file
This refactors the code necessary to generate the underlines and
srikeouts into a different file, cleaning up some of the code in the
process of doing so.

This also switches out some of the font metrics to use cell metrics
instead, which has the advantage of including additional information
like the offset specified in the configuration file.