Fix line metrics
The height of the line drawing for underline and strikethrough has been
incorrectly substracted from the Y position instead of adding it.
Leading to an incorrect Y position for underline and strikeout by
exactly the width of the line.

By simply adding half of the height instead of substracting it, the
position should be corrected and the distanace between the underline and
the baseline should be increased.