Keep log file logs themselves as they are.
It's going to be a pain to get the format of the logger in the context
of the log file itself. These logs are the first thing you see (or
last), so the visual impact is minimal. Also they aren't in the log
file, so the log file itself will be consistent.