Revert "Make the macOS bundle compatible osx 10.12"
This reverts commit 24533d2c5d86857dbb2f75f171430da7073ce65c.

First, this change does not allow to use custom MACOSX_DEPLOYMENT_TARGET.
Second, the original problem about minimum required version is not related to the alacritty configuration but looks like something specific to the old osx platform. Usually, the minimum required version is equal the version of used SDK if it is not provided explicitly. So the problem should be solved system-wide anyway.