Fix racing condition while displaying messages
Since the message was read from the message bar after the terminal lock
has already been freed, it was possible that the message would be
cleared while the renderer still expected it to be drawn, causing an
invalid state.

To resolve this, the message bar has been removed from the `Display`
struct and can now only be accessed through the Terminal.