Move message bar to crossbeam channels
Instead of using a locking Arc<Mutex<_>> for the message bar, it has
been moved to be owned exclusively by the `Term` with write-access to
the message queue using a crossbeam channel.

Explicit mouse click handling has also been added to the message bar,
which made it possible to add a simple `close` button in the form of
`[X]`.