Fix issues with multiple messages
With two messages in queue there were a few remaining bugs and crashes.
The size info wouldn't get updated properly when the bar height changed
with the second message, which would lead to a crash and it was possible
to close the second message's button with the release event after
clicking the first.

These remaining problems have been resolved and it is now possible to
have various different messages with different heights appear after each
other without crashes or bugs.