Render message background as rect
Instead of rendering the background of the message bar using the
terminal cells, it is now rendered as one big rectangle. This has the
advantage that it's possible to render it from edge to edge, making
additional padding at the sides less obvious and intrusive.

This also changes the behavior with the close button so it will not
overwrite existing cells anymore, instead it will force a line break if
the cells are already occupied. A single-cell padding has also been
added to the left of the button.