Remove selections when clearing screen partially
Automatically remove all selections when part of the screen is cleared.
This fixes issues in applications like `less -S` where a selection would
stay around after scrolling horizontally.

XTerm and URxvt both choose to always remove the selection, even if it's
outside of the cleared area, however VTE only clears the selection if
any part of it is inside the cleared area.

To keep things simple, Alacritty has adopted the behavior of XTerm and
URxvt to always clear selections.

This fixes #1644.