Add inherited configuration file support.
Adds two methods for including another yaml file inside of the config
file that is referenced by alacritty: `include PATH` and `include!  PATH`.

`include!` will throw an error if the path is not found while `include`
will only throw a warning.

It also substitutes environment variables which are referenced in the
file path to allow for a greater degree of customization via scripting.