Configurable window position
Adds the ability to specify window position through config file or command line
flag. Position is set early to avoid "teleporting" the window from its initial
to its configured position. The `position` field of `WindowConfig` is
wrapped in an `Option` because 0x0 is not the default window position -- we want
to let the window manager handle placement if position is not configured in
either the config or by passing the flag.