Underline URLs on hover
When a URL is hovered over, all characters that are recognized as part
of the URL will be underlined. After the cursor leaves the URL, the
previous hover state is restored.

This also changes the behavior when clicking an illegal character right
in front of a URL. Previously this would still launch the URL, but strip
the illegal character. Now these clicks are ignored to make sure there's
no mismatch between underline and legal URL click positions