Improve errors for missing shell file on Windows
Currently with WinPTY there is no error at all if the shell does not
exist and on ConPTY the error is cryptic 'assertion failed: success'.

This changes the ConPTY assertion to a human-readable error which
clearly states that the file could not be found.

Currently the WinPTY backend has not been fixed yet.

This fixes #2016.