Remove live-config-reload feature
This removes the live-config-reload feature to make the renderer
significantly more concise and maintainable.

As far as I'm aware, this feature was never really used by anyone since
there isn't much of an advantage to recompiling Alacritty yourself just
to make live config reload work.

This feature has also been broken for a long time, since the projection
is updated after the reload without considering the padding. Editing
with standard editors like vim also deletes the file and replaces it,
making it impossible to use them for live config reload.

Closes #1523.