Add NSSystemAdministrationUsageDescription to macOS manifest
This resolves permission failures when modifying root-owned files
and such, e.g., modifying /etc/hosts. It had been possible
to work around this by adding alacritty to the set of
applications with "Full Disk Access" in System Preferences. macOS
now opens a prompt to confirm permission with a new installation.

This fixes #2337.