Fix ncurses backspace key
Ncurses does some special handling for backspace, normally comparing it
to the KEY_BACKSPACE constant. However to make this work, the right
backspace escape sequence needs to be set in the terminfo.

XTerm has this set to ^H, but it always sends ^H. Since Alacritty
usually sends ^? instead, this needs to be specified to allow correct
interpretation for ncurses.