Allow setting both instance and class parts of WM_CLASS hint
When using a custom desktop entry to run a terminal application
within alacritty, gnome-shell will still treat the window as
belonging to the Alacritty desktop entry unless *both* instance
and class are changed to something else.  If both are changed,
the window will show up under the right icon in the dash.

Example:
  [Desktop]
  Name=Neovim
  Type=Application
  Icon=nvim
  Exec=alacritty --class nvim,nvim --title Neovim -e nvim

This will launch nvim in alacritty such that gnome-shell associates
the window with the Neovim icon.