Fix iterator start point errors
Since the display iterator does not include the initial cell in the
first `next()` or `prev()` col, the initial cell of the URL was not
correctly included.

Instead of having the field `cur` publicly accessibly on the
`DisplayIterator`, there are the two methods `DisplayIterator::point`
and `DisplayIterator::cell` for accessing the current element of the
iterator now.