WIP Implement damage tracking for damaged buffer swap
Track damage so that it can be reported to supported compositors.

Damage is tracked by increasing the size of a rect with the cursor
position every time the cursor is explicitly moved, as well as with the
final cursor position on readout. Doing it this way leaves Term::input
unaffected, minimizing overhead.