Resolve race between DPI change and resize events.
When the DPI changes, alacritty updates the terminal dimensions as
part of the handle_resize callback. However, the PhysicalSize with
which alacritty does the calculation can occasionally be the wrong
DPI, depending on the order in which the DPI change and resize
events are sent.

We're working around this race by asking the window for its current
DPI and size when we detect a DPI change.