Support saving window title in a stack
This commit adds the concept of a "title stack" to the terminal. Some
programs (e.g. vim) send control sequences `CSI 22 ; 0` (push title) and
`CSI 23 ; 0` (pop title). Clues on how to handles these were taken from
a couple of other terminal emulators (
[Gnome Terminal](https://github.com/GNOME/vte/commit/23c44f1f18b7985d9fba3205500d8873ecde51aa),
[kitty](https://github.com/kovidgoyal/kitty/commit/3067103b188c55d594918d1525d397cb523f8fea)).

The title stack is mainly just a history of previous titles.
Applications can push the current title onto the stack, and pop it back
off (setting the window title in the process).

To implement this, the Alacritty `Term` must have two new pieces of
state: `title` (the mirror of what is set in the Glutin window) and
`title_stack` (the history of titles if pushed onto). When control
sequence 22 is received, then a title is pushed, and 23 pops and sets
the Glutin window title.