Fix cases where config's "window" is not defined
In the config, if `window` is undefined, the derived `Default` for
the String `title` is used, which is an empty String. This was
unintended, and causes issues in gnome-shell (e.g. in the alt-tab
dialog) when the window title is an empty string.

This commit adds a manually implemented default for the `WindowConfig`,
it's the same as the derived `Default`, except for the `title`, which
will now always be "Alacritty" as originally intended.